Fix default scope and deleted scope in paranoia:
- Moved the class methods and queryable override to an actual default
  scope and deleted named scope.
- This exposed an issue with nested attributes and paranoid embedded
  document deletion, which was also fixed.
- [fix #2124]

+++ lib/mongoid/relations/embedded/many.rb
@@ -131,8 +131,8 @@ module Mongoid
         # @since 2.0.0.rc.1
         def delete(document)
           doc = target.delete_one(document)
-          _unscoped.delete_one(doc)
           if doc && !_binding?
+            _unscoped.delete_one(doc) unless doc.paranoid?
             if _assigning?
               if doc.paranoid?
                 doc.destroy(suppress: true)
